+++ src/components/Containers/Header/Header.js
@@ -1,100 +1,56 @@
 import React, {Fragment} from 'react';
-import {useLocation, useNavigate} from 'react-router-dom';
 import useBreakpoint from 'use-breakpoint';
 
-import {TrackEvent} from '../../../analytics';
 import {ReactComponent as StarkGateLogo} from '../../../assets/img/starkgate.svg';
-import {ReactComponent as BuyIcon} from '../../../assets/svg/tabs/buy.svg';
-import {ReactComponent as DiscordIcon} from '../../../assets/svg/tabs/discord.svg';
+import {ReactComponent as LiquidityIcon} from '../../../assets/svg/tabs/liquidity.svg';
 import {Breakpoint, ChainType} from '../../../enums';
-import {
-  useColors,
-  useConstants,
-  useEnvs,
-  useTracking,
-  useHeaderTranslation,
-  useBuyProviders
-} from '../../../hooks';
-import {useLogin} from '../../../providers/AppProvider';
+import {useColors, useEnvs, useHeaderTranslation, useLiquidityProviders} from '../../../hooks';
+import {useApp, useLogin} from '../../../providers/AppProvider';
 import {useMenu} from '../../../providers/MenuProvider';
 import {useIsL1, useIsL2} from '../../../providers/TransferProvider';
 import {useL1Wallet, useL2Wallet} from '../../../providers/WalletsProvider';
-import {openInNewTab, toClasses} from '../../../utils';
-import {Divider, Tab, WalletButton} from '../../UI';
+import {toClasses} from '../../../utils';
+import {Divider, Tab, WalletButton, BurgerMenu} from '../../UI';
 import styles from './Header.module.scss';
 
 export const Header = () => {
-  const {DISCORD_LINK_URL} = useConstants();
-  const [trackDiscordClick] = useTracking(TrackEvent.DISCORD_TAB_CLICK);
   const {supportedL1ChainId} = useEnvs();
-  const {tabBuyTxt, tabDiscordTxt, tabFaqTxt, tabTermsTxt, chainTxt} = useHeaderTranslation();
-  const navigate = useNavigate();
-  const {pathname} = useLocation();
+  const {tabLiquidityTxt, chainTxt} = useHeaderTranslation();
   const {showAccountMenu, showTransferMenu} = useMenu();
   const [, swapToL1] = useIsL1();
   const [, swapToL2] = useIsL2();
   const {account: l1Account, config: l1Config} = useL1Wallet();
   const {account: l2Account, config: l2Config} = useL2Wallet();
   const {breakpoint} = useBreakpoint(Breakpoint);
-  const {colorDiscord, colorWhiteOp50, colorGamma} = useColors();
+  const {colorGamma} = useColors();
+  const {navigateToRoute} = useApp();
   const {isLoggedIn} = useLogin();
-  const buyProviders = useBuyProviders();
-
-  const maybeNavigateToBridge = () => {
-    pathname !== '/' && navigate('/');
-  };
+  const liquidityProviders = useLiquidityProviders();
 
   const onL2WalletButtonClick = () => {
     swapToL2();
     showAccountMenu();
-    maybeNavigateToBridge();
+    navigateToRoute('/');
   };
 
   const onL1WalletButtonClick = () => {
     swapToL1();
     showAccountMenu();
-    maybeNavigateToBridge();
+    navigateToRoute('/');
   };
 
   const onLogoClick = () => {
     showTransferMenu();
-    maybeNavigateToBridge();
-  };
-
-  const onRouteTabClick = route => {
-    navigate(`/${route}`);
-  };
-
-  const onTabDiscordClick = () => {
-    trackDiscordClick();
-    openInNewTab(DISCORD_LINK_URL);
+    navigateToRoute('/');
   };
 
   const tabs = [
     {
-      color: colorDiscord,
-      icon: <DiscordIcon />,
-      text: tabDiscordTxt,
-      divider: true,
-      onClick: onTabDiscordClick
-    },
-    {
       color: colorGamma,
-      icon: <BuyIcon />,
-      text: tabBuyTxt,
-      disable: !buyProviders.length,
-      divider: true,
-      onClick: () => onRouteTabClick('buy')
-    },
-    {
-      color: colorWhiteOp50,
-      text: tabTermsTxt,
-      onClick: () => onRouteTabClick('terms')
-    },
-    {
-      color: colorWhiteOp50,
-      text: tabFaqTxt,
-      onClick: () => onRouteTabClick('faq')
+      icon: <LiquidityIcon />,
+      text: tabLiquidityTxt,
+      disable: !liquidityProviders.length,
+      onClick: () => navigateToRoute('liquidity')
     }
   ];
 
@@ -104,7 +60,7 @@ export const Header = () => {
         !tab.disable && (
           <Fragment key={index}>
             <Tab colorBorder={tab.color} icon={tab.icon} text={tab.text} onClick={tab.onClick} />
-            {tab.divider && <Divider />}
+            {index !== tabs.length - 1 && <Divider />}
           </Fragment>
         )
       );
@@ -138,6 +94,7 @@ export const Header = () => {
             />
           </>
         )}
+        <BurgerMenu />
       </div>
     </div>
   );