+++ src/components/UI/index.js
@@ -1,6 +1,8 @@
 export * from './BackButton/BackButton';
 export * from './Badge/Badge';
 export * from './Button/Button';
+export * from './Modal';
+export * from './Toast';
 export * from './Circle/Circle';
 export * from './CryptoLogo/CryptoLogo';
 export * from './Icon/Icon';
@@ -36,6 +38,6 @@ export * from './MultiChoiceErrorMessage/MultiChoiceErrorMessage';
 export * from './MultiChoiceMenu/MultiChoiceMenu';
 export * from './MultiChoiceItem/MultiChoiceItem';
 export * from './Divider/Divider';
+export * from './BurgerMenu/BurgerMenu';
+export * from './BurgerMenuItem/BurgerMenuItem';
 export * from './CollapseExpand/CollapseExpand';
-export * from './Modal';
-export * from './Toast';