+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.5.0-dev.7](https://github.com/starkware-libs/starkgate-frontend/compare/v1.5.0-dev.6...v1.5.0-dev.7) (2022-05-23)
+
+
+### 🧩 Features
+
+* add LayerSwap to liquidity providers ([#201](https://github.com/starkware-libs/starkgate-frontend/issues/201)) ([4f8aa97](https://github.com/starkware-libs/starkgate-frontend/commit/4f8aa97cc0b8e8a159e0b2ba8935422751cf3226))
+
+## [1.5.0-dev.6](https://github.com/starkware-libs/starkgate-frontend/compare/v1.5.0-dev.5...v1.5.0-dev.6) (2022-05-23)
+
+
+### 🧩 Features
+
+* move the discord tab to a side button ([#199](https://github.com/starkware-libs/starkgate-frontend/issues/199)) ([0570a92](https://github.com/starkware-libs/starkgate-frontend/commit/0570a922b90a7069e7b71c14bbc75b707fb34c54))
+
 ## [1.5.0-dev.5](https://github.com/starkware-libs/starkgate-frontend/compare/v1.5.0-dev.4...v1.5.0-dev.5) (2022-05-22)